Added Async/Await To Submission
Typically API requests are going to be asynchronous, i.e. the user is
going to dispatch a request to the API, wait for the server to respond
and then once the server has responded, render the information. There
are many ways to handle asynchronous requests with React, such as
chaining on a promise to the API request function's body, but an easier
way is to utilize the async await syntax:
  - Define the function as async by prepending the async keyword to the
    function definition
  - Define the request as await by prepending the await keyword to
    the API request
  - Assign the API request with the await keyword to a constant variable
    called response
  - For now, console log out the response body (specifically with
    the data.results accessors called on the object to get the data
    we are looking to render)
  - We receive an array of objects that we can now manipulate in our
    rendering