Added Keys and Alt Text To Images
React throws an error when the key prop is not passed to a list of
objects because it formally does not know how to order items that are
added to or removed from a list without the keys. I.e., if we had a list
of tasks that were to be performed in a specific order, 1, 2, 3, 4, etc
and we wanted to add a new item to this list, unless it has a key
specifying what order it is supposed to be performed in attached to it
then it will not be able to be added to the list. Keys can be passed
as props to the rendered instances of the images, using the id property.

React also throws a warning when alt text is not implemented. Alt text
is used for accessibility but also in case the images fail to render.
Each image contains a description property that we can pass as a
variable to the alt property of the image JSX tag.

Finally, we initially are writing our properties as image.id,
image.description, image.urls.regular. Because we want to use the DRY
best practice, we destructure out the image portion of the prop and in
our map function pass in an object of id, description, and urls,
allowing us to remove the image accessor on each of these properties.